Trigger reflow on document.elementsFromPoint
