+++ tests/wpt/mozilla/meta/MANIFEST.json
@@ -6341,6 +6341,18 @@
      {}
     ]
    ],
+   "mozilla/document_elementsFromPoint.html": [
+    [
+     "/_mozilla/mozilla/document_elementsFromPoint.html",
+     [
+      [
+       "/_mozilla/mozilla/document_elementsFromPoint_ref.html",
+       "=="
+      ]
+     ],
+     {}
+    ]
+   ],
    "mozilla/iframe/resize_after_load.html": [
     [
      "/_mozilla/mozilla/iframe/resize_after_load.html",
@@ -9384,6 +9396,11 @@
     ]
    ],
    "mozilla/details_ui_opened_ref.html": [
+    [
+     {}
+    ]
+   ],
+   "mozilla/document_elementsFromPoint_ref.html": [
     [
      {}
     ]
@@ -25136,6 +25153,14 @@
   "mozilla/document_defaultView.html": [
    "12eef74e9ba1a62b0302df79633cab5c48f4977f",
    "testharness"
+  ],
+  "mozilla/document_elementsFromPoint.html": [
+   "2349e1d102552a46bb8e3dd36c0b3396e0d6530b",
+   "reftest"
+  ],
+  "mozilla/document_elementsFromPoint_ref.html": [
+   "4d1a1d0acc16bcb123781cbed47184ff308f3097",
+   "support"
   ],
   "mozilla/document_getElementById.html": [
    "ae5270afde878c8fbe9b4a89c5c3f8ff609220dc",
