+++ tests/wpt/mozilla/tests/mozilla/document_elementsFromPoint.html
@@ -1,0 +1,7 @@
+<!doctype html>
+<meta charset="utf-8">
+<title></title>
+<link rel="match" href="document_elementsFromPoint_ref.html">
+<script>
+    var elements = document.elementsFromPoint(10, 10);
+</script>
