+++ tests/wpt/mozilla/tests/mozilla/document_elementsFromPoint_ref.html
@@ -1,0 +1,3 @@
+<!doctype html>
+<meta charset="utf-8">
+<title></title>
